requirements.txt
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-WTF==1.1.1
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
python-decouple==3.7
python-dotenv==0.21.1
six==1.16.0
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.12.0
